functional safety check.
git-svn-id: svn://192.168.1.10/public/CoffeeMud@16404 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ com/planet_ink/coffee_mud/core/CMStrings.java
@@ -1359,7 +1359,7 @@ public class CMStrings
 		if((numValues==0)||(firstIndex<0))
 			return;
 		final int valueLen=(numValues<=10)?1:Integer.toString(numValues).length();
-		long safety=100;
+		short safety=100;
 		for(int i=firstIndex;(i<str.length()-(1+valueLen));i++)
 		{
 			if((str.charAt(i)=='@')
